Iter9 metrics dump to file commit 8 refactoring
